% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImpACC.R
\name{varImpACC}
\alias{varImpACC}
\title{varImpACC}
\usage{
varImpACC(object, mincriterion = 0, conditional = FALSE, threshold = 0.2,
  nperm = 1, OOB = TRUE, pre1.0_0 = conditional)
}
\arguments{
\item{object}{}

\item{mincriterion}{}

\item{conditional}{}

\item{threshold}{}

\item{nperm}{}

\item{OOB}{}

\item{pre1.0_0}{}
}
\value{
vector with computed permutation importance for each variable
}
\description{
varImpACC
}
\examples{
data(iris)
iris2 = iris
iris2$Species = factor(iris$Species == "versicolor")
iris.cf <- cforest(Species ~ ., data = iris2,control = cforest_unbiased(mtry = 2, ntree = 50))
set.seed(123)
a = varImpACC(object = iris.cf)

}
