% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varImpAUC.R
\name{varImpAUC}
\alias{varImpAUC}
\title{varImpAUC}
\usage{
varImpAUC(object, mincriterion = 0, conditional = FALSE, threshold = 0.2,
  nperm = 1, OOB = TRUE, pre1.0_0 = conditional, method = c("ovo",
  "ova"))
}
\arguments{
\item{object}{}

\item{mincriterion}{}

\item{conditional}{}

\item{threshold}{}

\item{nperm}{}

\item{OOB}{}

\item{pre1.0_0}{}

\item{method}{}
}
\value{
vector with computed permutation importance for each variable
}
\description{
varImpAUC
}
\examples{

# multiclass case
data(iris)
iris.cf <- cforest(Species ~ ., data = iris, control = cforest_unbiased(mtry = 2, ntree = 50))
set.seed(123)
a = varImpAUC(object = iris.cf, method = "ovo")
set.seed(123)
b = varImpAUC(object = iris.cf, method = "ova") 
}
